fix(repo): fix pr sugs
metriport/metriport-internal#1040

Ref: ENG-479
