+++ packages/api/src/routes/medical/patient-root.ts
@@ -6,7 +6,6 @@ import {
   PaginatedResponse,
   stringToBoolean,
 } from "@metriport/shared";
-import { buildDayjs } from "@metriport/shared/common/date";
 import dayjs from "dayjs";
 import duration from "dayjs/plugin/duration";
 import { Request, Response } from "express";
@@ -39,8 +38,7 @@ import {
 import { fromCreateResponseToDto, PatientImportDto } from "./dtos/patient-import";
 import { dtoFromModel, PatientDTO } from "./dtos/patientDTO";
 import { schemaCreateToPatientData, schemaDemographicsToPatientData } from "./schemas/patient";
-import { createTcmEncounter } from "../../command/medical/tcm-encounter/create-tcm-encounter";
-import { sample } from "lodash";
+import { createSampleTcmEncounters } from "./create-sample-tcm-encounter";
 
 dayjs.extend(duration);
 
@@ -96,49 +94,7 @@ router.post(
     });
 
     if (!Config.isProdEnv()) {
-      const facilityNames = [
-        "Riverside Regional Medical Center",
-        "St. Catherine's General Hospital",
-        "Pine Valley Medical Center",
-        "Heritage Community Hospital",
-        "Westfield Memorial Hospital",
-      ];
-
-      // Create 3 TCM encounters with specific dates and events
-      const twentyDaysAgo = buildDayjs().subtract(20, "day");
-      const tenDaysAgo = buildDayjs().subtract(10, "day");
-      const yesterday = buildDayjs().subtract(1, "day");
-
-      const encounters = [
-        {
-          latestEvent: "Discharged" as const,
-          admitTime: twentyDaysAgo,
-          dischargeTime: twentyDaysAgo.add(22, "hour"),
-          clinicalInformation: {},
-        },
-        {
-          latestEvent: "Discharged" as const,
-          admitTime: tenDaysAgo,
-          dischargeTime: tenDaysAgo.add(35, "hour"),
-          clinicalInformation: {},
-        },
-        { latestEvent: "Admitted" as const, admitTime: yesterday, clinicalInformation: {} },
-      ];
-
-      await Promise.all(
-        encounters.map(async enc => {
-          return await createTcmEncounter({
-            cxId,
-            patientId: patient.id,
-            class: sample(["Inpatient", "Emergency"]) ?? "",
-            facilityName: sample(facilityNames) ?? "",
-            latestEvent: enc.latestEvent,
-            admitTime: enc.admitTime.toDate(),
-            dischargeTime: enc.dischargeTime?.toDate(),
-            clinicalInformation: enc.clinicalInformation,
-          });
-        })
-      );
+      await createSampleTcmEncounters(cxId, patient.id);
     }
 
     return res.status(httpStatus.CREATED).json(dtoFromModel(patient));
